requirements.txt
discord.py==2.4.0
Requests==2.32.3
yt_dlp==2024.7.9
